Langchain Learning/4. Tools/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
distro==1.9.0
dnspython==2.7.0
frozenlist==1.5.0
greenlet==3.1.1
groq==0.37.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.21
langchain-core==0.3.83
langchain-groq==0.3.8
langchain-text-splitters==0.3.7
langsmith==0.3.45
loguru==0.7.3
motor==3.7.0
multidict==6.2.0
orjson==3.10.15
packaging==24.2
propcache==0.3.0
pydantic==2.10.6
pydantic_core==2.27.2
pymongo==4.11.3
python-dotenv==1.2.2
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
sniffio==1.3.1
SQLAlchemy==2.0.39
tenacity==9.0.0
typing_extensions==4.12.2
urllib3==2.3.0
uuid_utils==0.14.1
win32_setctime==1.2.0
yarl==1.18.3
zstandard==0.23.0
